Yeet Composer, Add scope, Yeet World & Commands sugar, Add EntityCommands sugar
